chart adjustments
